Changed player & lyrics screens to swipeable modals Make dark mode background color compatible with AMOLED screens
