+++ src/themes/dark.js
@@ -1,11 +1,11 @@
 export const dark = {
 	foreground: '#1ED760',
 	foreground2: 'white',
-	background: 'black',
+	background: '#121212',
 	elevatedBG: '#1D222D',
 	contrast: 'white',
 	contrastTrans: 'rgba(255, 255, 255, ',
 	fgTrans: 'rgba(30, 215, 96, ',
-	bgTrans: 'rgba(0, 0, 0, ',
+	bgTrans: 'rgba(18, 18, 18, ',
 	current: 'dark'
 };
